+++ docs/CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 - Log warning when `amethyst_test::WaitForLoad` has not completed in 10 seconds. ([#1984])
 - Derive `Copy` and `PartialEq` for `amethyst::renderer::resources::Tint`. ([#2033])
 - Derive `Hash` for `amethyst::input::{Button, ControllerButton, ScrollDirection}`. ([#2041])
+- Added Trans::Replace, Trans::NewStack, and Trans::Sequence to the State Machine Transitions. ([#2067],[#2071])
 - Add rendy/gfx-backend log verbosity through configuration. ([#1652]) ([#2048])
 
 ### Changed
@@ -76,9 +77,12 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 [#2029]: https://github.com/amethyst/amethyst/pull/2029
 [#2033]: https://github.com/amethyst/amethyst/pull/2033
 [#2041]: https://github.com/amethyst/amethyst/pull/2041
+[#2067]: https://github.com/amethyst/amethyst/issue/2067
 [#2048]: https://github.com/amethyst/amethyst/pull/2048
+[#2071]: https://github.com/amethyst/amethyst/pull/2071
 [#2070]: https://github.com/amethyst/amethyst/pull/2070
 [#2079]: https://github.com/amethyst/amethyst/pull/2079
+
 
 
 ## [0.13.3] - 2019-10-4